+++ tests/Status.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Celema\Boiler\Tests;
+
+enum Status
+{
+	case Active;
+	case Inactive;
+}